medical-project-backend/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
apispec==6.8.1
    # via
    #   -r requirements.in
    #   apispec-pydantic-plugin
apispec-pydantic-plugin==0.6.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
bcrypt==4.2.1
click==8.1.8
    # via
    #   flask
    #   typer
    #   uvicorn
python-dotenv==1.0.1
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-pydantic
flask-bcrypt==1.0.1
flask-pydantic==0.12.0
flask-cors==5.0.0
flask-jwt-extended==4.7.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.26.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
packaging==24.2
    # via
    #   apispec
    #   marshmallow
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   apispec-pydantic-plugin
    #   flask-pydantic
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pyotp==2.9.0
pytest==8.3.4
requests==2.32.3
rich==13.9.4
    # via typer
shellingham==1.5.4
    # via typer
toml==0.10.2
    # via -r requirements.in
typer==0.15.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   typer
uvicorn==0.34.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
